move updateSoapAssessment function to patient route
